+++ src/ezdxf/math/ellipse.py
@@ -1,19 +1,18 @@
 # Copyright (c) 2020, Manfred Moitzi
 # License: MIT License
-from typing import TYPE_CHECKING, Iterable, Dict
+from typing import TYPE_CHECKING, Iterable, Dict, cast
 import math
 from collections import namedtuple
 from .vector import Vector, NULLVEC, X_AXIS, Z_AXIS
 from .matrix44 import Matrix44
-from .ucs import OCS
+from .ucs import OCS, UCS
 from .construct2d import rytz_axis_construction, enclosing_angles, linspace
-from ezdxf.math import bspline
 
 Params = namedtuple('Params', 'center major_axis minor_axis extrusion ratio start end')
 pi2 = math.pi / 2
 
 if TYPE_CHECKING:
-    from ezdxf.eztypes import Vertex
+    from ezdxf.eztypes import Vertex, BaseLayout, Ellipse
 
 QUARTER_PARAMS = [0, math.pi * .5, math.pi, math.pi * 1.5]
 HALF_PI = math.pi / 2.0
@@ -167,18 +166,20 @@ class ConstructionEllipse:
         self.start_param = (start_param - HALF_PI) % math.tau
         self.end_param = (end_param - HALF_PI) % math.tau
 
-    def spline(self, num: int = 16) -> bspline.BSpline:
-        """ Returns a curve approximation as spline with `num` control points. """
-        fit_points = list(self.vertices(self.params(num)))
-        count = len(fit_points)
-        degree = 2
-        order = degree + 1
-        t_vector = list(bspline.uniform_t_vector(fit_points))
-        knots = list(bspline.control_frame_knots(count - 1, degree, t_vector))
-        control_points = bspline.global_curve_interpolation(fit_points, degree, t_vector, knots)
-        spline = bspline.BSpline(control_points, order=order, knots=knots)
-        spline.t_array = t_vector
-        return spline
+    def add_to_layout(self, layout: 'BaseLayout', dxfattribs: dict = None) -> 'Ellipse':
+        """
+        Add ellipse as DXF :class:`~ezdxf.entities.Ellipse` entity to a layout.
+
+        Args:
+            layout: destination layout as :class:`~ezdxf.layouts.BaseLayout` object
+            dxfattribs: additional DXF attributes for DXF :class:`~ezdxf.entities.Ellipse` entity
+
+        """
+        from ezdxf.entities import Ellipse
+        dxfattribs = dxfattribs or dict()
+        dxfattribs.update(self.dxfattribs())
+        e = Ellipse.new(dxfattribs=dxfattribs, doc=layout.doc)
+        return cast(Ellipse, layout.add_entity(e))
 
 
 def transform(params: Params, m: Matrix44) -> Params: