+++ NEWS.md
@@ -28,6 +28,8 @@ Version 0.13 - dev
 - NEW: `Hatch.set_pattern_angle()` to set rotation angle of pattern definition
 - NEW: `DXFGraphic.unlink_from_layout()` to unlink entity from associated layout
 - NEW: `ezdxf.math.ConstructionEllipse()`
+- NEW: `BSpline.from_arc()` returns a B-Spline approximation for a `ConstructionArc()`
+- NEW: `BSpline.from_ellipse()` returns a B-Spline approximation for a `ConstructionEllipse()`
 - CHANGE: `Hatch` full support for rotated patterns.
 - CHANGE: `Hatch.set_pattern_definition()` added argument `angle` for pattern rotation. 
 - CHANGE: renamed 2D `ConstructionXXX.move()` methods to `translate()`